% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank.R
\docType{data}
\name{drugbank}
\alias{drugbank}
\title{DrugBank vocabulary}
\format{
A data frame with columns:

A data frame with 53940 rows and 10 variables
\describe{
\item{id}{DrugBank identifier.}
\item{name}{character for drug's common name.}
\item{synonyms}{string expression for alternative names, delimited by the pipe character.}
}
}
\source{
DrugBank Open Data dataset,
\url{https://go.drugbank.com/releases/latest#open-data}.
}
\usage{
drugbank
}
\description{
DrugBank identifiers, names, and synonyms.
}
\keyword{datasets}
