% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_temp_data.R
\docType{data}
\name{example_temp_data}
\alias{example_temp_data}
\title{Example temperature data}
\format{
A data frame with 144 rows and 3 variables:
\describe{
  \item{nest}{Nest ID}
  \item{date}{Date (YYYY-MM-DD)}
  \item{time}{Time (HH:MM:SS)}
  \item{temp}{Temperature (degrees centigrade)}
}
}
\source{
Hypothetical data of two nests over 1 or 2 days
}
\usage{
data(example_temp_data)
}
\description{
Example data set for two nests over one or two days.
}
\keyword{datasets}
